Finishes the features for part B1.
You can now add, edit, and delete different terms
from the main page of the application.
I utilized the "add term" page for the editing of each term as well.
That way I didn't not have to create redundant code.
This also includes new files to use the weak message references so I can
notify other parts of the app when certain methods are finished.